Weather/requirements.txt
schedule>=0.5.0
requests>=2.21.0
paho-mqtt>=1.4.0